Moved method to own file
