+++ githubpoxy.go
@@ -20,14 +20,3 @@ func (ghp GitHubProvider) Policies() ([]Policy, error) {
 func (ghp GitHubProvider) DocumentDownload(downloadKey string) (io.ReadCloser, error) {
 	return nil, nil
 }
-
-func LoadCredentialsFromEnvironment() (username string, password string, err error) {
-	username = os.Getenv("username")
-	password = os.Getenv("password")
-
-	if username != "" && password != "" {
-		return username, password, nil
-	}
-
-	return "", "", errors.New("Error loading credentials from environment. (username, password)")
-}
